+++ test/utest-online/testlinkapi_online_test.py
@@ -353,10 +353,25 @@ class TestLinkAPIOnlineTestCase(unittest.TestCase):
         with self.assertRaisesRegex(TLResponseError, '3000.*40000711'):
             self.client.deleteTestPlan(40000711)
 
-    def test_getTestCaseKeywords_unknownID(self):
+    # test might fail during Travis test, cause used TestLink demo application
+    # represents  still a 1.9.13 dev state from 26/12/14
+    # the keyword add method are added later and will be changed with 1.9.14
+    # the interface (see TL Mantis Task 6934)        
+    @unittest.expectedFailure
+    def test_addTestCaseKeywords_unknownID(self):
         with self.assertRaisesRegex(TLResponseError, '5040.*TC-40000712'):
             self.client.addTestCaseKeywords('TC-40000712', 
                                             ['KeyWord01', 'KeyWord03']) 
+
+    # test might fail during Travis test, cause used TestLink demo application
+    # represents  still a 1.9.13 dev state from 26/12/14
+    # the keyword remove method are added later and will be changed with 1.9.14
+    # the interface (see TL Mantis Task 6907)      
+    @unittest.expectedFailure
+    def test_removeTestCaseKeywords_unknownID(self):
+        with self.assertRaisesRegex(TLResponseError, '5040.*TC-40000712'):
+            self.client.removeTestCaseKeywords('TC-40000712', 
+                                            ['KeyWord01']) 
              
         
 if __name__ == "__main__":